Adds executable accessor
