+++ apps/base/modules/Api/ExecutorApi.ts
@@ -4,7 +4,7 @@ import {
   Executor
 } from "../modules";
 
-export class ExecutableApi {
+export class ExecutorApi {
 
   constructor(
     private executor: Executor,
@@ -20,9 +20,13 @@ export class ExecutableApi {
      * method: POST
      */
     this.serverCommunicator.post(ApiPaths.RUN_EXECUTABLE, (req: any, resp: any) => {
-      const executable = req.body.executable;
-      const data = req.body.data;
-      this.executor.runExecutable(executable, data)
+      const username = req.params.username;
+      const cluster = req.params.cluster;
+      const environment = req.params.environment;
+      const exe = req.params.exe;
+      const name = req.params.name;
+      const data = req.body;
+      this.executor.runExecutable(username, cluster, environment, exe, name, data)
       .then((result: any) => {
         resp.status(200).send({result: result});
       });
