+++ apps/base/modules/Executable/ExecutableFactory.ts
@@ -3,7 +3,6 @@ import { Query } from "./Query";
 import { Graph } from "./Graph";
 import { ShellCommunicator } from "../Communicator/ShellCommunicator";
 import { FileSystem } from "../FileSystem/FileSystem";
-import { DatabaseCommunicator } from "../Communicator/DatabaseCommunicator";
 import { FileSystemFactory } from "../FileSystem/FileSystemFactory";
 import { EnvironmentClient } from "../Client/EnvironmentClient";
 
@@ -13,7 +12,6 @@ export class ExecutableFactory {
     private fileSystemFactory: FileSystemFactory,
     private fileSystem: FileSystem,
     private shell: ShellCommunicator,
-    private database: DatabaseCommunicator,
     private environmentClient: EnvironmentClient) {
 
   }
@@ -33,14 +31,15 @@ export class ExecutableFactory {
   }
 
   query(data: any): Query {
-    return new Query(
-      data.username,
-      data.name,
-      data.visibility,
-      data.text,
-      this.database,
-      this.environmentClient
-    )
+    // return new Query(
+    //   data.username,
+    //   data.name,
+    //   data.visibility,
+    //   data.text,
+    //   this.database,
+    //   this.environmentClient
+    // )
+    return undefined
   }
 
   graph(data: any): Graph {
