+++ images/client/Dockerfile
@@ -1,2 +1,2 @@
-FROM icanplayguitar94/iam:base-f71673374d245e911c8dd80a42e888ad92cfb105
+FROM icanplayguitar94/iam:base-minikube
 COPY ./src public/dist
